tidy up and changeset
