+++ .changeset/many-deers-brake.md
@@ -5,3 +5,5 @@
 callout: Added info tone and compact variant
 
 callout: Added support for custom icons
+
+box: Add support for more colours in background, foreground and border props
