+++ packages/react/src/callout/Callout.tsx
@@ -1,8 +1,13 @@
 import { ElementType, PropsWithChildren, ReactNode } from 'react';
 import { Flex } from '../flex';
-import { InfoFilledIcon } from '../icon';
 import { Stack } from '../stack';
 import { CalloutTitle } from './CalloutTitle';
+import {
+	CalloutTone,
+	calloutToneMap,
+	CalloutVariant,
+	calloutVariantMap,
+} from './utils';
 
 export type CalloutProps = PropsWithChildren<{
 	as?: ElementType;
@@ -15,9 +20,9 @@ export type CalloutProps = PropsWithChildren<{
 	/** Title will appear in bold */
 	title?: string;
 	/** Tone will change the background color */
-	tone?: keyof typeof calloutToneMap;
+	tone?: CalloutTone;
 	/** Variant will change the padding and gap */
-	variant?: keyof typeof calloutVariantMap;
+	variant?: CalloutVariant;
 }>;
 
 export const Callout = ({
@@ -58,33 +63,3 @@ export const Callout = ({
 		</Flex>
 	);
 };
-
-export const calloutToneMap = {
-	neutral: {
-		background: 'shade',
-		border: 'border',
-		icon: undefined,
-	},
-	info: {
-		background: 'info',
-		border: 'info',
-		icon: (
-			<InfoFilledIcon
-				aria-hidden="false"
-				aria-label="Information"
-				color="info"
-			/>
-		),
-	},
-} as const;
-
-export const calloutVariantMap = {
-	regular: {
-		gap: 1,
-		padding: 1.5,
-	},
-	compact: {
-		gap: 0.5,
-		padding: 1,
-	},
-} as const;
